+++ R/ArraySpeciesBySize-class.R
@@ -159,6 +159,8 @@ print.summary.ArraySpeciesBySize <- function(x, ...) {
 #' @param wlim A numeric vector of length two providing lower and upper limits
 #'   for the weight (x) axis. Use `NA` to refer to the existing minimum or
 #'   maximum. Only applies to `ArraySpeciesBySize`.
+#' @param size_axis Whether to plot size as weight (`"w"`, default) or length
+#'   (`"l"`), using the allometric weight-length relationship.
 #' @param ylim A numeric vector of length two providing lower and upper limits
 #'   for the value (y) axis. Use `NA` to refer to the existing minimum or
 #'   maximum.
@@ -180,14 +182,18 @@ print.summary.ArraySpeciesBySize <- function(x, ...) {
 #' \donttest{
 #' plot(getEncounter(NS_params))
 #' plot(getFeedingLevel(NS_params), species = c("Cod", "Herring"))
+#' plot(getPredMort(NS_params), species = c("Cod", "Herring"),
+#'      size_axis = "l")
 #' }
 plot.ArraySpeciesBySize <- function(x, species = NULL,
                             all.sizes = FALSE, highlight = NULL,
                             return_data = FALSE, log_x = TRUE, log_y = FALSE,
                             log = NULL,
                             wlim = c(NA, NA), ylim = c(NA, NA),
+                            size_axis = c("w", "l"),
                             total = FALSE, background = TRUE,
                             y_ticks = 6, ...) {
+    size_axis <- plot_size_axis(size_axis)
     log_axes <- parsePlotLog(log, log_x = log_x, log_y = log_y)
     log_x <- log_axes$log_x
     log_y <- log_axes$log_y
@@ -198,6 +204,7 @@ plot.ArraySpeciesBySize <- function(x, species = NULL,
     plot_dat <- prepare_ArraySpeciesBySize_plot_data(
         x, species = species, all.sizes = all.sizes, wlim = wlim,
         total = total, background = background)
+    plot_dat <- convert_plot_size_axis(plot_dat, params, size_axis)
 
     if (return_data) return(plot_dat)
 
@@ -206,11 +213,13 @@ plot.ArraySpeciesBySize <- function(x, species = NULL,
         y_label <- paste0(value_name, " [", units_str, "]")
     }
 
-    plotDataFrame(plot_dat, params, xlab = "Size [g]", ylab = y_label,
+    plotDataFrame(plot_dat, params, xlab = plot_size_xlab(size_axis),
+                  ylab = y_label,
                   xtrans = if (log_x) "log10" else "identity",
                   ytrans = if (log_y) "log10" else "identity",
-                  xlim = wlim, ylim = ylim, highlight = highlight,
-                  y_ticks = y_ticks, legend_var = "Legend")
+                  xlim = plot_size_xlim(wlim, size_axis), ylim = ylim,
+                  highlight = highlight, y_ticks = y_ticks,
+                  legend_var = "Legend")
 }
 
 parsePlotLog <- function(log, log_x = FALSE, log_y = FALSE) {
@@ -258,10 +267,12 @@ plot2.ArraySpeciesBySize <- function(x, y, name1 = "First", name2 = "Second",
                                      species = NULL, all.sizes = FALSE,
                                      log_x = TRUE, log_y = FALSE, log = NULL,
                                      wlim = c(NA, NA), ylim = c(NA, NA),
+                                     size_axis = c("w", "l"),
                                      total = FALSE, background = TRUE,
                                      y_ticks = 6, ...) {
     check_plot2_compatible(x, y, "ArraySpeciesBySize")
     compare_array_metadata(x, y)
+    size_axis <- plot_size_axis(size_axis)
     log_axes <- parsePlotLog(log, log_x = log_x, log_y = log_y)
     log_x <- log_axes$log_x
     log_y <- log_axes$log_y
@@ -277,11 +288,12 @@ plot2.ArraySpeciesBySize <- function(x, y, name1 = "First", name2 = "Second",
 
     plotComparisonDataFrame(plot_dat1, plot_dat2, params,
                             name1 = name1, name2 = name2,
-                            xlab = "Size [g]", ylab = y_label,
+                            xlab = plot_size_xlab(size_axis), ylab = y_label,
                             xtrans = if (log_x) "log10" else "identity",
                             ytrans = if (log_y) "log10" else "identity",
-                            xlim = wlim, ylim = ylim,
-                            y_ticks = y_ticks, legend_var = "Legend")
+                            xlim = plot_size_xlim(wlim, size_axis), ylim = ylim,
+                            y_ticks = y_ticks, legend_var = "Legend",
+                            size_axis = size_axis)
 }
 
 #' Plot the relative difference between two mizer array objects
@@ -316,10 +328,12 @@ plotRelative.ArraySpeciesBySize <- function(x, y, species = NULL,
                                             log_x = TRUE,
                                             wlim = c(NA, NA),
                                             ylim = c(NA, NA),
+                                            size_axis = c("w", "l"),
                                             total = FALSE,
                                             background = TRUE, ...) {
     check_plot2_compatible(x, y, "ArraySpeciesBySize")
     compare_array_metadata(x, y)
+    size_axis <- plot_size_axis(size_axis)
     params <- attr(x, "params")
     plot_dat1 <- prepare_ArraySpeciesBySize_plot_data(
         x, species = species, all.sizes = all.sizes, wlim = wlim,
@@ -329,9 +343,10 @@ plotRelative.ArraySpeciesBySize <- function(x, y, species = NULL,
         total = total, background = background)
 
     plotRelativeDataFrame(plot_dat1, plot_dat2, params,
-                          xlab = "Size [g]",
+                          xlab = plot_size_xlab(size_axis),
                           xtrans = if (log_x) "log10" else "identity",
-                          xlim = wlim, ylim = ylim, legend_var = "Legend")
+                          xlim = plot_size_xlim(wlim, size_axis), ylim = ylim,
+                          legend_var = "Legend", size_axis = size_axis)
 }
 
 check_plot2_compatible <- function(x, y, class) {
@@ -405,6 +420,7 @@ addPlot <- function(plot, x, ...) {
 addPlot.ArraySpeciesBySize <- function(plot, x, species = NULL,
                                        all.sizes = FALSE,
                                        wlim = c(NA, NA),
+                                       size_axis = c("w", "l"),
                                        total = FALSE,
                                        background = TRUE,
                                        colour = NULL,
@@ -419,15 +435,19 @@ addPlot.ArraySpeciesBySize <- function(plot, x, species = NULL,
                 is.number(alpha),
                 alpha >= 0,
                 alpha <= 1)
+    size_axis <- plot_size_axis(size_axis)
 
     plot <- deep_copy(plot)
     plot_dat <- prepare_ArraySpeciesBySize_plot_data(
         x, species = species, all.sizes = all.sizes, wlim = wlim,
         total = total, background = background)
-    check_addPlot_compatible(plot, x_var = "w", y_var = "value",
+    params <- attr(x, "params")
+    plot_dat <- convert_plot_size_axis(plot_dat, params, size_axis)
+    x_var <- plot_size_x_var(size_axis)
+    check_addPlot_compatible(plot, x_var = x_var, y_var = "value",
                              units = attr(x, "units"))
 
-    mapping <- aes(x = .data[["w"]], y = .data[["value"]],
+    mapping <- aes(x = .data[[x_var]], y = .data[["value"]],
                    group = .data[["Species"]])
     if (is.null(colour)) {
         mapping$colour <- rlang::quo(.data[["Legend"]])